refactor: decouple metadata linking from motionFinalized to utils
