+++ src/handlers/motions/motionCreated/handlers/simpleDecision.ts
@@ -1,9 +1,10 @@
 import { BigNumber } from 'ethers';
 
 import { ContractEvent, motionNameMapping } from '~types';
+import { SimpleTransactionDescription } from '~utils/actions/parseAction';
 import { getColonyDecisionId } from '~utils/decisions';
 
-import { SimpleTransactionDescription, createMotionInDB } from '../helpers';
+import { createMotionInDB } from '../helpers';
 
 export const handleSimpleDecisionMotion = async (
   event: ContractEvent,
